backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic[email]==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.26.2
pytesseract==0.3.10
pillow==10.1.0
python-socketio==5.10.0
aiohttp==3.9.1
pytest==7.4.3
httpx==0.25.2 